:recycle: refactor: update cli package
