+++ packages/serendipity-core/package.json
@@ -5,7 +5,8 @@
   "author": "yuzhanglong <yuzl1123@163.com>",
   "homepage": "https://github.com/yuzhanglong/serendipity/tree/master/packages/serendipity-core#readme",
   "license": "MIT",
-  "main": "bin/index.js",
+  "main": "lib/index.js",
+  "module": "esm/index.js",
   "scripts": {
     "start": "tsc -w",
     "lib:cjs": "tsc -p tsconfig.json --target ESNext --module commonjs --outDir lib",
