+++ packages/serendipity-core/src/types/pluginMeta.ts
